changed old audio files with better; can load and save user high score; created round window; created fly away window with fly away script that is updated on not enough bullets; ducks will now stay in screen bounds until they have to fly away; created end of round window and end of round script that currently have some end game logic, based on ducks spawned; fixed bugs related to input events that were caused by monkey actions, when there isn't InputProcessor assigned in the Input class

+++ src/core/graphics/game_object_renderer.cpp
@@ -19,10 +19,10 @@ void GameObjectRenderer::draw(const float& dt) {
 void GameObjectRenderer::drawObject(EntityAnimated* e, const float& dt) {
 	e->processAnimation(dt);
 	if(e->isFacingLeft()){
-		e->getCurrentFrame().draw(e->getX(), e->getY(), e->getWidth(), e->getHeight(), SDL_FLIP_HORIZONTAL);
+		e->getCurrentFrame().draw((int) e->getX(), (int) e->getY(), (int) e->getWidth(), (int) e->getHeight(), SDL_FLIP_HORIZONTAL);
 	}
 	else{
-		e->getCurrentFrame().draw(e->getX(), e->getY(), e->getWidth(), e->getHeight());
+		e->getCurrentFrame().draw((int) e->getX(), (int) e->getY(), (int) e->getWidth(), (int) e->getHeight());
 	}
 }
